+++ tests/test_base.py
@@ -1,0 +1,30 @@
+# -*- coding: utf-8 -*-
+
+import os
+
+import unittest
+# import seismograph
+import urlparse
+import time
+
+from selenium.webdriver import DesiredCapabilities, Remote
+from selenium.webdriver.support.ui import WebDriverWait
+from selenium.common.exceptions import TimeoutException
+from selenium.webdriver.common.action_chains import ActionChains
+
+class Page(object):
+    BASE_URL = 'https://ok.ru/'
+    PATH = ''
+
+    def __init__(self, driver):
+        self.driver = driver
+
+    def open(self):
+        url = urlparse.urljoin(self.BASE_URL, self.PATH)
+        self.driver.get(url)
+        self.driver.maximize_window()
+
+        
+class Component(object):
+    def __init__(self, driver):
+        self.driver = driver
